api/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
python-jose[cryptography]
python-multipart
numpy<2.0
scikit-learn==1.5.1
pandas
prometheus-client
prometheus-fastapi-instrumentator
requests
Pillow
matplotlib
rapidfuzz
scipy
scikit-surprise
psycopg2
passlib
